fix(models): 渠道分组编辑器可选中仅存在于模型库的模型
系统默认分组的模型选择器列不出只有 model_configs 行、没有运行时来源的
模型，导致这类模型无法被加入分组白名单。两个原因叠加：

- effectiveModels 未把 AvailabilityFilterOptions 传给 withScopedModelConfigRows，
  ignore_group_allowed_models 管不到模型库行这条支路。分组自身的 allowed-models
  于是过滤了它自己被编辑的候选集：模型要先在分组里才能被加进分组。
- modelOwnerScope 仅按 match 规则匹配账号，default 池没有 match 块，
  解析不出任何 owner，模型库行永远进不了系统默认分组。

编辑器路径改为忽略 allowed-models（plaza/catalog 仍照常执行白名单），
default 池成员改为向 auth 包查询（新增 ChannelGroupsForAuthForTenant），
避免在管理端复制一份会与路由漂移的归属判定。

availability.go 因此超过 1200 行门禁，按规范拆分而非放宽门禁：owner 解析
移入 availability_owner_scope.go，主文件 1188 → 993 行并锁定基线。

本地验证：新增两个回归测试（先红后绿）、./scripts/ci-pr.sh 完整通过。
