+++ docs/internal-review/backend-structure-allowlist.json
@@ -26,7 +26,7 @@
       },
       {
         "path": "internal/management/modelcatalog/availability.go",
-        "max_lines": 1188
+        "max_lines": 993
       },
       {
         "path": "internal/management/settings/providers/provider_keys_extended.go",
